1.1.6
fix：修复NumberFormat.format格式化Float数据偶发精度丢失的bug

+++ stepper/src/main/java/com/rouxinpai/stepper/Stepper.kt
@@ -447,7 +447,7 @@ class Stepper @JvmOverloads constructor(
 
     private fun setTextAndMoveSelection(value: Float?) {
         with(mBinding.etInput) {
-            setText(value?.let { mFormatter.format(it) })
+            setText(value?.let { mFormatter.format(it.toBigDecimal()) })
             setSelection(text?.length ?: 0)
         }
     }
